otaclient: integrate new log_setting module into main
