+++ otaclient/app/log_setting.py
@@ -14,8 +14,19 @@
 
 import logging
 import os
+import yaml
 
 from .configs import config as cfg
+
+
+# NOTE: EcuInfo imports this log_util so independent get_ecu_id are required.
+def _get_ecu_id():
+    try:
+        with open(cfg.ECU_INFO_FILE) as f:
+            ecu_info = yaml.load(f, Loader=yaml.SafeLoader)
+            return ecu_info["ecu_id"]
+    except Exception:
+        return "autoware"
 
 
 def configure_logging(loglevel: int, *, http_logging_url: str):
