+++ otaclient/app/main.py
@@ -14,19 +14,20 @@
 
 
 import asyncio
+import logging
 import os
 import sys
 from pathlib import Path
 
 from otaclient import __version__  # type: ignore
-from . import log_util
 from .common import read_str_from_file, write_str_to_file_sync
 from .configs import config as cfg, EXTRA_VERSION_FILE, OTACLIENT_LOCK_FILE
+from .log_setting import configure_logging, _get_ecu_id
 from .ota_client_service import launch_otaclient_grpc_server
 
-logger = log_util.get_logger(
-    __name__, cfg.LOG_LEVEL_TABLE.get(__name__, cfg.DEFAULT_LOG_LEVEL)
-)
+# configure logging before any code being executed
+configure_logging(loglevel=cfg.DEFAULT_LOG_LEVEL, http_logging_url=_get_ecu_id())
+logger = logging.getLogger(__name__)
 
 
 def _check_other_otaclient():
